Cellular automata simulation of traffic flow
This is my second year project. It is a cellular automata model of traffic flow. There are three files: a single lane system, a multilane system, and the multilane system with maximum velocities applied for different vehicles. More comments and explanations are to be added to these for better context at a later date.